Create Multiple_Of_Five_Grades.py
This code calculates the difference between each grade in a given list and the next multiple of five of each one, if this difference is less than 3, than this grade will be round to the next multiple of five, if the difference is highter, the number will not be rounded. Grades smaller than 38 will not be rounded independent if the difference is highter than 3 or not.